Refine Problem 10. Regular Expression Matching

+++ src/RegularExpressionMatching.cpp
@@ -5,16 +5,15 @@
 using namespace std;
 
 bool RegularExpressionMatching::isMatch(string s, string p) {
-    int sl = s.size();
-    int pl = p.size();
+    size_t sl = s.size();
+    size_t pl = p.size();
     vector<vector<bool>> dp(sl + 1, vector<bool>(pl + 1, false));
-    // if p == null, then s and p matches iff s == null
+
     dp[0][0] = true;
 
     for (int i = 1; i <= sl; i++)
         dp[i][0] = false;
 
-    // if s == null, then s and p matches iff p conforms to (.*)*
     for (int j = 1; j <= pl; j++)
         dp[0][j] = j > 1 && p[j - 1] == '*' && dp[0][j - 2];
